Fix neutralize target bug introducedi wth SLG
The neutralize target objective map was updated to support sparks. This broke evac in several ways:

- If the last player evac'd is carrying the VIP, the script incorrectly kills the VIP. The script
  was intended to kill the VIP if it was abandoned with no players left capable of carrying it.
- You can cheese the mission by knocking out the VIP then evacing everyone. It'll kill the VIP, but
  you'll still win.

This patch to the kismet fixes the first problem. When the last soldier is evac'd and the VIP is
unconscious or bleeding out, first check to see if it's still on the map before killing it. If it's
not, it's already been evac'd or is in the process of being evac'd with this action.

The fix requires kismet changes, so it will conflict with any other mod that modifies the neutralize
target mission definition.